gst-helper: Update to files to git d4d7218fbd6c902ef0fb4265e4d1bb7bbd8f2702

+++ libs/gst-helper/include/tcamprop_system/tcamprop_provider_funcbased.h
@@ -27,12 +27,12 @@ namespace tcamprop_system
 
         using set_func_button = std::function<std::error_code()>;
 
-        void register_menu( const property_info& nfo, const std::vector<std::string_view>& menu_entries, int default_menu_entry, set_value_func<int> set, get_value_func<int> get, get_flags_func get_flags );
-        void register_menu( const property_info& nfo, const std::vector<std::string_view>& menu_entries, int default_menu_entry, set_value_func<int> set, get_value_func<int> get, prop_flags def_flags = prop_flags::def_flags );
+        void register_menu( const property_info& nfo, const std::vector<std::string_view>& menu_entries, int64_t default_menu_entry, set_value_func<int64_t> set, get_value_func<int64_t> get, get_flags_func get_flags );
+        void register_menu( const property_info& nfo, const std::vector<std::string_view>& menu_entries, int64_t default_menu_entry, set_value_func<int64_t> set, get_value_func<int64_t> get, prop_flags def_flags = prop_flags::def_flags );
         void register_double( const property_info& nfo, const prop_range_real& range, set_value_func<double> set, get_value_func<double> get, get_flags_func get_flags );
         void register_double( const property_info& nfo, const prop_range_real& range, set_value_func<double> set, get_value_func<double> get, prop_flags def_flags = prop_flags::def_flags );
-        void register_integer( const property_info& nfo, const prop_range_integer& range, set_value_func<int> set, get_value_func<int> get, get_flags_func get_flags );
-        void register_integer( const property_info& nfo, const prop_range_integer& range, set_value_func<int> set, get_value_func<int> get, prop_flags def_flags = prop_flags::def_flags );
+        void register_integer( const property_info& nfo, const prop_range_integer& range, set_value_func<int64_t> set, get_value_func<int64_t> get, get_flags_func get_flags );
+        void register_integer( const property_info& nfo, const prop_range_integer& range, set_value_func<int64_t> set, get_value_func<int64_t> get, prop_flags def_flags = prop_flags::def_flags );
         void register_boolean( const property_info& nfo, bool def, set_value_func<bool> set, get_value_func<bool> get, get_flags_func get_flags );
         void register_boolean( const property_info& nfo, bool def, set_value_func<bool> set, get_value_func<bool> get, prop_flags def_flags = prop_flags::def_flags );
         void register_button( const property_info& dsc, set_func_button set, get_flags_func get_flags );
